refactor(agents): All agents now use locking by default
