+++ orca-migration/src/main/kotlin/com/netflix/spinnaker/orca/pipeline/persistence/migration/PipelineMigrationAgent.kt
@@ -36,7 +36,7 @@ class PipelineMigrationAgent(
 
   private val log = LoggerFactory.getLogger(javaClass)
 
-  override fun startPolling() {
+  override fun tick() {
     val previouslyMigratedPipelineIds = dualExecutionRepository.primary.retrieveAllExecutionIds(PIPELINE)
 
     val executionCriteria = ExecutionRepository.ExecutionCriteria().apply {
